pyproject.toml
[tool.poetry]
name = "youtube-video-chat"
version = "0.1.0"
description = "Simple chatbot that answers questions about YouTube video transcripts"
authors = ["Francesco Amedeo Fontana <franz.fonta@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.29.0"
openai = "^1.3.7"
youtube-transcript-api = "^0.6.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
